requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.1
